feat(python): represent Float64 as a plain Python float
Completes the numeric representation change: F# `float` now compiles to a plain
Python `float` instead of a pyo3 `Float64` wrapper, and plain `float` is its
runtime tag. Every float64 literal in the generated test corpus becomes bare --
`float64(` goes from 2363 occurrences to 0.

This half is simpler than Int32: a Python `float` *is* an IEEE double, so `+ - *`,
overflow-to-inf, NaN comparison, banker's rounding and signed zero are already
bit-identical to .NET and need no normalization at all. Only three things diverge,
and all were verified against .NET:

- `1.0 / 0.0` raises ZeroDivisionError in Python where .NET yields +/-inf, and
  `0.0 / 0.0` should be nan. Routed through `op_division_float64`, with a fast path
  that keeps `x / 2.0` a bare operator when the divisor is a non-zero literal.
- Python's float `%` takes the sign of the divisor where .NET takes the dividend's:
  `-5.0 % 3.0` is 1.0 in Python and -2.0 in .NET. Routed through
  `op_remainder_float64`, which is `math.fmod` plus a nan guard for the cases where
  fmod raises but .NET returns nan.
- String interpolation: .NET renders a whole double without a trailing ".0", so
  `$"Circle {r}"` with r = 5.0 gives "Circle 5". Python's `str` gives "5.0", so
  interpolated floats route through the existing `exceptions.to_string`.

`:? float` compiles to an exact `type(x) is float` check, which also keeps it
distinct from the Float32 wrapper -- that is not a `float` subclass.

Fixed several places where the wrapper leaked back into code that now expects
plain floats, none of which the tests would have caught:

- The module-level float functions in floats.rs (`sqrt`, `floor`, `atan2`, ...)
  accepted plain floats but returned `Float64`. They now return plain f64.
- `double.py` rebound `float64` to the wrapper class, shadowing the coercion
  imported from `.core`, so `sign` returned a wrapper.
- `double.inf`/`nan`/`negative_inf` were wrapper constants.
- `util.round` called `value.round()`, which a plain float does not have. It now
  uses the same scaling algorithm as the wrapper; `builtins.round` rounds half to
  even, as .NET does.
- Native array `sum` and the sort comparer boxed Float64 results.

Verified: 2428 Python tests, 275 fable-library-py tests. Pyright unchanged at 37
errors -- Float64 introduced none.

+++ src/fable-library-py/fable_library/array_.py
@@ -15,7 +15,7 @@ UInt32ArrayCons = array.FSharpCons[uint32]("UInt32")
 Int64ArrayCons = array.FSharpCons[int64]("Int64")
 UInt64ArrayCons = array.FSharpCons[uint64]("UInt64")
 Float32ArrayCons = array.FSharpCons[float32]("Float32")
-Float64ArrayCons = array.FSharpCons[float64]("Float64")
+Float64ArrayCons = array.FSharpCons[float]("Float64")
 BoolArrayCons = array.FSharpCons[bool]("Bool")
 
 # Typed array classes (use these for isinstance checks and type annotations).
